+++ src/main/java/me/fullpotato/badlandscaves/SupernaturalPowers/DescensionStage/DetectionDecrease.java
@@ -3,7 +3,6 @@ package me.fullpotato.badlandscaves.SupernaturalPowers.DescensionStage;
 import me.fullpotato.badlandscaves.BadlandsCaves;
 import me.fullpotato.badlandscaves.Util.PlayerScore;
 import org.bukkit.entity.Player;
-import org.bukkit.metadata.FixedMetadataValue;
 import org.bukkit.scheduler.BukkitRunnable;
 
 public class DetectionDecrease extends BukkitRunnable {
@@ -19,7 +18,7 @@ public class DetectionDecrease extends BukkitRunnable {
             if (in_descension == 2) {
                 double detection = (PlayerScore.DESCENSION_DETECT.hasScore(plugin, player)) ? ((double) PlayerScore.DESCENSION_DETECT.getScore(plugin, player)) : 0;
                 if (detection > 0) {
-                    int detect_cooldown = (int) PlayerScore.DESCENSION_DETECT_COOLDOWN.getScore(plugin, player);
+                    double detect_cooldown = (double) PlayerScore.DESCENSION_DETECT_COOLDOWN.getScore(plugin, player);
                     if (detect_cooldown <= 0) {
                         detection -= 0.5;
                         PlayerScore.DESCENSION_DETECT.setScore(plugin, player, detection);
