+++ src/main/java/me/fullpotato/badlandscaves/SupernaturalPowers/Spells/Runnables/ManaRegen.java
@@ -3,7 +3,6 @@ package me.fullpotato.badlandscaves.SupernaturalPowers.Spells.Runnables;
 import me.fullpotato.badlandscaves.BadlandsCaves;
 import me.fullpotato.badlandscaves.Util.PlayerScore;
 import org.bukkit.entity.Player;
-import org.bukkit.metadata.FixedMetadataValue;
 import org.bukkit.scheduler.BukkitRunnable;
 
 public class ManaRegen extends BukkitRunnable {
@@ -22,7 +21,7 @@ public class ManaRegen extends BukkitRunnable {
             double max_mana = ((double) PlayerScore.MAX_MANA.getScore(plugin, player));
             if (Mana >= max_mana) return;
 
-            int thirst = (int) PlayerScore.THIRST.getScore(plugin, player);
+            double thirst = (double) PlayerScore.THIRST.getScore(plugin, player);
             if (thirst < 30) return;
 
             int mana_regen_delay_timer = ((int) PlayerScore.MANA_REGEN_DELAY_TIMER.getScore(plugin, player));
